requirements.txt
altair==3.2.0
ansi2html==1.6.0
attrs==19.3.0
biopython==1.76
brotli==1.2.0
certifi==2020.6.20
charset-normalizer==2.0.12
click==7.1.2
contextvars==2.4
cycler==0.10.0
dash==2.15.0
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
dash_ag_grid==32.3.2
dataclasses==0.8
DateTime==4.3
engineering-notation==0.10.0
Flask==1.1.4
Flask-Compress==1.15
fonttools==4.27.1
future==1.0.0
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.15.3
packaging==20.9
pandas==0.23.4
Pillow==8.4.0
pkgutil_resolve_name==1.3.10
plotly==5.18.0
primers==0.5.4
psutil==5.9.0
pyfastx==0.8.4
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.3
requests==2.27.1
retrying==1.3.3
scikit-learn==0.20.0
scipy==1.1.0
seqfold==0.7.17
six==1.17.0
tenacity==7.0.0
tinydb==4.5.2
tk-tools==0.16.0
toolz==0.11.2
tqdm==4.64.0
urllib3==1.26.12
Werkzeug==1.0.1
XlsxWriter==3.2.2
zope.interface==5.4.0
zstandard==0.20.0
